refactor: move filter type declarations to root of filter package
