+++ internals/api/dto/processor_dto.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"image"
 
-	"dougdomingos.com/image-filters/filters/types"
+	"dougdomingos.com/image-filters/filters"
 )
 
 // ProcessorRequestDTO represents the expected parameters for a parsed
@@ -26,7 +26,7 @@ type ProcessorRequestDTO struct {
 
 	// Pipeline is the requested filter pipeline that matches the "filter"
 	// parameter from the request.
-	Pipeline types.FilterPipeline
+	Pipeline filters.FilterPipeline
 }
 
 // ProcessorResponseDTO represents the JSON for processor HTTP responses.
