+++ internals/api/services/benchmark.go
@@ -10,7 +10,6 @@ import (
 
 	"dougdomingos.com/image-filters/engines"
 	"dougdomingos.com/image-filters/filters"
-	"dougdomingos.com/image-filters/filters/types"
 	"dougdomingos.com/image-filters/internals/api/dto"
 )
 
@@ -67,7 +66,7 @@ func parseBenchmarkRequest(r *http.Request) (*dto.BenchmarkRequestDTO, int, stri
 // benchmarkPipeline applies a filter pipeline to an image and returns the
 // processing duration in milliseconds. The caller may specify if the pipeline
 // should run concurrently or not.
-func benchmarkPipeline(pipeline types.FilterPipeline, img image.RGBA) int64 {
+func benchmarkPipeline(pipeline filters.FilterPipeline, img image.RGBA) int64 {
 	start := time.Now()
 	engines.ApplyFilterPipeline(&img, &pipeline)
 	duration := time.Since(start)
